+++ ArtGalleryUI/src/app/features/checkout/address-list/address-list.component.ts
@@ -1,5 +1,5 @@
 import { AsyncPipe } from '@angular/common';
-import { Component, OnDestroy, OnInit } from '@angular/core';
+import { Component, EventEmitter, Input, OnDestroy, OnInit, Output } from '@angular/core';
 import { ActivatedRoute, Router, RouterLink } from '@angular/router';
 import { AddressList } from '../models/address-list.model';
 import { Observable, Subscription } from 'rxjs';
@@ -15,6 +15,9 @@ import { AddressService } from '../services/address.service';
 export class AddressListComponent implements OnInit, OnDestroy{
   userId:any= '';
   model?: AddressList[];
+  @Input() parentComponent = '';
+  @Output() addressSelectEvent = new EventEmitter<AddressList>();
+  selectedAddress? : AddressList;
   private getAddressesByUserIdSubscription?: Subscription;
   private paramsSubscription?: Subscription;
   private deleteAddressSubscription?: Subscription;
@@ -40,23 +43,31 @@ export class AddressListComponent implements OnInit, OnDestroy{
     });
   }
 
+  onAddressSelect(address:AddressList ){
+    this.selectedAddress = address;
+    this.addressSelectEvent.emit(address);
+  }
+
   onDeleteClick(id: string) {
-    this.deleteAddressSubscription = this.addressService
+    if(confirm("Are you to delete the address?")){
+      this.deleteAddressSubscription = this.addressService
       .deleteAddress(id)
       .subscribe({
         next: (response) => {
           this.router
-            .navigateByUrl('/', { skipLocationChange: true })
-            .then(() => {
-              this.router.navigate([`user/address/${this.userId}`]);
-            });
+          .navigateByUrl('/', { skipLocationChange: true })
+          .then(() => {
+            this.router.navigate([`user/address/${this.userId}`]);
+          });
         },
       });
-  }
-
+    }
+    }
+    
   ngOnDestroy(): void {
     this.paramsSubscription?.unsubscribe();
     this.getAddressesByUserIdSubscription?.unsubscribe();
+    this.deleteAddressSubscription?.unsubscribe();
   }
 
 }
